# doc/openapi.yaml
openapi: 3.0.3
info:
  title: NextRise API
  version: 0.1.0
  description: |
    Draft OpenAPI specification for the NextRise backend (Next.js route handlers).
    This file was generated from the current source tree. Many response / schema
    details are inferred and should be refined once domain models are finalized.

  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Local dev server
  - url: https://api.nexrise.example.com
    description: Production (placeholder)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nr_session   # cookieOptions.name in code (adjust if different)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    UserPublic:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
      required: [id, email]
    AuthUserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserPublic'
...
security:
  - cookieAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email & password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie is set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          description: Bad request, missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
      responses:
        '201':
          description: User created
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/incubator/news:
    get:
      tags: [News]
      summary: List news (paginated)
      parameters:
        - in: query
          name: skip
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, default: 100 }
      responses:
        '200': { description: OK }
        '500': { description: Error }
...